feat(backups): implement backup script
instead of copying the entire contents of the named volume via a crazy
process of mounting it to a different container etc. etc. we just use
`docker compose cp` to copy what we need out of the containers.

we do this because there's a ton of extraneous crap in the volumes (i2pd
has a bunch of files with peer data, for example) it's easier to just
copy the five files we need and be done with it.